requirements.txt
accelerate==0.24.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
appnope==0.1.3
astor==0.8.1
asttokens==2.4.1
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bitsandbytes==0.41.2.post2
blessed==1.20.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.17
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.3
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.4.2
docutils==0.20.1
editor==1.6.5
executing==2.0.1
fastapi==0.109.2
filelock==3.13.1
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.49.0
frappe-manager==0.8.3
frozenlist==1.4.0
fsspec==2024.2.0
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
googleapis-common-protos==1.61.0
grpcio==1.59.2
h11==0.14.0
html2image==2.0.4.3
httpcore==1.0.3
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.19.3
humanfriendly==10.0
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
inquirer==3.2.1
ipykernel==6.26.0
ipython==8.17.2
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.0
jupyter_core==5.5.0
keyring==24.3.0
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.0.349
langchain-community==0.0.1
langchain-core==0.0.13
langsmith==0.0.69
litellm==1.18.8
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
more-itertools==10.1.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nh3==0.2.14
nodeenv==1.8.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.16.2
open-interpreter==0.2.0
openai==1.9.0
opencv-python==4.9.0.80
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.8.0
pillow==10.2.0
pkginfo==1.9.6
platformdirs==4.0.0
posthog==3.3.2
prisma==0.12.0
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.8
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
PyGetWindow==0.0.9
Pygments==2.16.1
pymongo==4.6.1
PyMsgBox==1.0.9
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Quartz==10.1
pyparsing==3.1.1
PyPDF2==3.0.1
pyperclip==1.8.2
PyPika==0.48.9
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytweening==1.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.1
readchar==4.0.5
readme-renderer==42.0
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
rpds-py==0.13.2
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
rubicon-objc==0.4.7
runs==1.2.0
scipy==1.12.0
seaborn==0.13.2
selenium==4.16.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.36.3
streamlit==1.29.0
sympy==1.12
tenacity==8.2.3
thop==0.1.1.post2209072238
tiktoken==0.4.0
tokenizers==0.15.0
tokentrim==0.1.13
toml==0.10.2
tomlkit==0.12.3
toolz==0.12.0
torch==2.2.0
torchvision==0.17.0
tornado==6.3.3
tqdm==4.66.2
traitlets==5.13.0
trio==0.23.2
trio-websocket==0.11.1
twine==4.0.2
tymongo @ file:///Users/majidraimi/Developer/Projects/Packages/tymongo/dist/tymongo-0.0.4-py3-none-any.whl#sha256=c1bd74856644403f9934adca62ff160974bd0b8ea430232604d6ff10016c9eb9
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
ultralytics==8.1.15
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
validators==0.22.0
watchfiles==0.21.0
wcwidth==0.2.10
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.0.1
wget==3.2
wrapt==1.16.0
wsproto==1.2.0
xmod==1.8.1
yarl==1.9.4
zipp==3.17.0
